requirements.txt
fastapi
pydantic
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain_community
python-multipart
bcrypt
